Remove XAA 'acceleration'.
We have a lot of stale code to 'accelerate' certain operations
through XAA. However, in practice, this acceleration is
completely unhelpful because whenever we encounter an
unaccelerated action, we have to sync and flush the fifo
which kills any performance gain.

As such, the virtual hardware doesn't even advertise these
acceleration capabilities anymore, so the code is completely
unusued.

In addition, XAA is on the way out, so there's no point
leaving in dead code which will have to go in the fairly
near future.

The one operation we can meaningfully accelerate is a
front-only fill and when we get around to implementing
that, we'll use EXA instead.