Fix a race condition when switching to a multihead fullscreen mode.
Due to implemtation details on the host side, switching to a multihead
fullscreen mode means multiple mode changes along the way. However,
because rrSetScreenConfig returns before the mode change completes,
we can enter a race, where a subsequent SetDisplayTopology overrides
the current pending one, leading to a mode switch where the resolution
and topology are no longer consistent.

So, we fix this by ignoring any attempts to set the pending topology if
one is already pending.