Video Overlay: Add XV_AUTOPAINT_COLORKEY capability to the Xv extension of Linux guest driver.
Applications use XV_AUTOPAINT_COLORKEY capability to let the driver handle the painting
of colorkey.

Real Player assumes this capability as ON by default and hence prior to this patch videos
didn't play with Real player.

With this patch:
a) If AUTOPAINT_COLORKEY is SET then the driver handles the painting of the colorkey.
Painting is done on the 1st frame and from there on only if the video frame moves.

b) Adds UYVY fourcc format to keep it consistent with Windows SVGA driver and the host backends.

c) Changes the default colorkey from a pungent GREEN to a darker shade.
Real Player sets this colorkey and it looks better when the video is moved around.