Video Overlay: Handle clipping correctly with AutoPaint colorkey.
We recently added XV_AUTOPAINT_COLORKEY attribute to the X video driver
to enable applications like Real player that rely on this attribute to
display video frames in Linux guest.

When this attribute is SET, we paint the colorkey on 1st frame and from
there on only when the video is moved.

This introduced a bug 305202 with clipping.

Consider a case when the video is playing, obscure the player window
with another window, without moving the player window, move the other
window away. The part of the window that was obscured didn't display
the video.

With this patch instead of relying on the target rectangle, we use
the clipBoxes supplied in every frame.