fix a cursor refcounting bug, leading to segfaults
this is similar to what xf86_use_hw_cursor() does, which is replaced by
vmwareUseHWCursor (otherwise, the refcount could reach zero and hence the
cursor deallocated while xf86CursorEnableDisableFBAccess() could still bring
it back to life from the saved cursor).
It is probably insane to do refcounting here, but this needs a xserver fix,
and even if that's fixed this fix here shouldn't hurt (though would be
unnecessary).