vmwlegacy: Fix segfault and setup of initial mode.
The fact that modes were added without names caused segfaults on older
X servers. At least up to and including Xserver 1.4.

Also, for some reason, at least Xserver 1.4 insists on setting the first mode
in the modelist, even if we set another mode as the current one.

Work around this by inserting a new mode with the current screen dimensions,
and add that modename last to the array of display requested mode names.
This means that if none of the previous mode names are found, we will at least
find the newly inserted mode.

Also, if there are no requested mode names at all, the driver previously
chose the largest mode that fit the timings. Now we will, in that case,
always select the newly inserted mode and thus not change resolution unless
specified.

Also add an option to not add this default mode. The option
"AddDefaultMode" is true by default.

Finally when we restore registers at exit and VT switch, make sure we
reprogram the initial width, height and bpp for the next time we start a
server.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>