vmwlegacy: Fix server termination due to pitch inconsistency.
Don't require a 32*32 bit pitch alignment when validating modes, since the
requested virtual pitch (pVMWARE->maxWidth*bpp) must be a multiple of that
alignment. If not, the server will terminate with a cryptic error message.

This is only for validating modes. The driver will adjust the pitch to the
host requirement when a mode is set anyway, and hopefully the host won't
require a pitch it doesn't support.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>