vmwgfx: Don't dirty the front buffer as hw when doing a dri2copyregion from it
When copying from a dri2 buffer we usually dirty it as hw, since dri has
been rendering to it, and there can only be hw contents in the buffer.

However for the real front, X has already done the dirty work for us.

Also remove a glxWaitX() debug message.

This should fix piglit read-front.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>